+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- feat(branch-prefix): implement support for customisable branch prefix input parameter (default: `chore/quarto-extensions`).
+
 ## 0.0.1 (2025-10-19)
 
 - Initial release of Quarto Extensions Updater.